🌱 Refactor: split sqlite.go monolith (3321 lines) into 8 domain files
Extract SQLiteStore methods into domain-specific files:

| File                    | Domain                          | Lines |
|-------------------------|---------------------------------|-------|
| sqlite.go (core)        | Struct, constructor, migrate     |   642 |
| sqlite_users.go         | Users + onboarding              |   253 |
| sqlite_dashboards.go    | Dashboards + cards + swaps      |   667 |
| sqlite_features.go      | Feature requests + notifications |   411 |
| sqlite_gpu.go           | GPU reservations + utilization  |   553 |
| sqlite_rewards.go       | User rewards + token usage      |   500 |
| sqlite_auth.go          | Token revocation + OAuth state  |   144 |
| sqlite_ops.go           | Cluster groups + audit + timeline|   228 |

No behavioral changes — all methods remain on *SQLiteStore in the
same package. Shared helpers (nullString, boolToInt, rollbackConn)
stay in sqlite.go.

Fixes #10385
